fix(rand): use safe modulo operation with byte conversion
Fix gosec security warning by explicitly converting byte to int before
modulo operation: charSet[int(v)%lsc] instead of charSet[v%byte(lsc)].
Prevents potential implicit type conversions.

+++ generate/rand.go
@@ -8,12 +8,14 @@ import (
 func stringWithCharset(strSize int) (string, error) {
 	charSet := "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
 	lsc := len(charSet)
+
 	bys := make([]byte, strSize)
 	if _, err := rand.Read(bys); err != nil {
 		return "", fmt.Errorf("error in reading random data %s", err)
 	}
+
 	for k, v := range bys {
-		bys[k] = charSet[v%byte(lsc)]
+		bys[k] = charSet[int(v)%lsc]
 	}
 
 	return string(bys), nil
